feat: Update labourSchema and labourModel
The commit updates the `labourSchema` and `labourModel` files. In the `labourSchema.ts` file, the `assignedHours` field is changed to have a minimum value of 2 and a maximum value of 18. Additionally, a new export named `labourUpdateSchema` is added with fields similar to `labourSchema`. In the `labourModel.ts` file, the `assignedHours` field is updated to have a minimum value of 2 and a maximum value of 18.